refactor: SuccessAlert, PrescribingContext
