+++ src/modules/common/components/card/ExaminationCard/hooks/useExaminationCard.js
@@ -2,7 +2,7 @@ import { doc, serverTimestamp, setDoc } from "firebase/firestore"
 import { useState } from "react"
 import { useTranslation } from "react-i18next"
 import { db } from "../../../../../../config/firebase"
-import SuccessfulAlert, { ConfirmAlert, ErrorAlert } from "../../../../../../config/sweetAlert2"
+import { ConfirmAlert, ErrorAlert } from "../../../../../../config/sweetAlert2"
 import { APP_ENV, STATUS_BOOKING_CONFIRMED, TOAST_SUCCESS } from "../../../../../../lib/constants"
 import { fetchSendEmailConfirmExamination } from "../services"
 import createToastMessage from "../../../../../../lib/utils/createToastMessage"
